refactor(user): simplify utils location and minor changes
