+++ app/api/routes/login.py
@@ -27,7 +27,7 @@ def login(
         raise HTTPException(status_code=400, detail="Inactive user")
     access_token_expires = timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
     return Token(
-        access_token=security.create_acess_token(
+        access_token=security.create_access_token(
             user.id, expires_delta=access_token_expires
         )
     )
