+++ app/api/utils.py
@@ -4,8 +4,6 @@ from email.mime.text import MIMEText
 from app.core.config import settings
 
 
-
-
 def send_email(to_email: str, subject: str, content: str):
     msg = MIMEMultipart()
     msg["From"] = settings.SMTP_USER
@@ -20,6 +18,7 @@ def send_email(to_email: str, subject: str, content: str):
         server.login(settings.SMTP_USER, settings.SMTP_PASSWORD)
         server.sendmail(settings.SMTP_USER, to_email, msg.as_string())
         server.quit()
-        print("Email enviado com sucesso AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA!")
+        print("Email enviado com sucesso!")
     except Exception as e:
-        print(f"Erro ao enviar email: {e} BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB")
+        print(f"Erro ao enviar email: {e}")
+
