+++ app/core/security.py
@@ -5,17 +5,19 @@ from datetime import datetime, timedelta, timezone
 import jwt
 
 
-pwd_context = CryptContext(schemes = ['bcrypt'], deprecated = 'auto')
+pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+
 
 def hash_password(password: str) -> str:
     return pwd_context.hash(password)
 
 
-def create_acess_token(subject: str | Any, expires_delta: timedelta) -> str:
+def create_access_token(subject: str | Any, expires_delta: timedelta) -> str:
     expire = datetime.now(timezone.utc) + expires_delta
     to_encode = {"exp": expire, "sub": str(subject)}
     encode_jwt = jwt.encode(to_encode, settings.SECRET_KEY)
     return encode_jwt
+
 
 ALGORITHM = "HS256"
